Refactor imports to use the new Text component from "@/components/ui/text" across multiple screens and components, enhancing consistency in the codebase. Removed the deprecated index.ts file from components directory.
