+++ app/(tabs)/explore.tsx
@@ -1,7 +1,7 @@
 import { ScrollView, StyleSheet } from "react-native";
 
-import { Text } from "@/components";
 import { ThemedView } from "@/components/ThemedView";
+import { Text } from "@/components/ui/text";
 
 export default function ExploreScreen() {
   return (
