+++ app/(tabs)/profile.tsx
@@ -1,4 +1,3 @@
-import { Text } from "@/components";
 import { ThemedView } from "@/components/ThemedView";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Button } from "@/components/ui/button";
@@ -35,12 +34,6 @@ export default function ProfileScreen() {
     ? `${user.firstName || ""} ${user.lastName || ""}`.trim()
     : "Loading...";
   const avatarUrl = user?.avatarUrl;
-  const initials = fullName
-    .split(" ")
-    .map((n) => n[0])
-    .join("")
-    .toUpperCase()
-    .slice(0, 2);
 
   return (
     <ThemedView className="flex-1">
@@ -60,9 +53,15 @@ export default function ProfileScreen() {
               <ActivityIndicator color="white" />
             </AvatarFallback>
           </Avatar>
-          <Text className="mb-1 text-center text-2xl font-semibold">{fullName}</Text>
-          <Text className="mb-3 text-center text-base">{user?.location || ""}</Text>
-          <Text className="mb-6 text-center text-base px-5">{user?.bio || ""}</Text>
+          <Text className="mb-1 text-center text-2xl font-semibold">
+            {fullName}
+          </Text>
+          <Text className="mb-3 text-center text-base">
+            {user?.location || ""}
+          </Text>
+          <Text className="mb-6 text-center text-base px-5">
+            {user?.bio || ""}
+          </Text>
           <Button
             variant="outline"
             className="w-full"
