+++ components/HelloWave.tsx
@@ -8,7 +8,7 @@ import Animated, {
   withTiming,
 } from "react-native-reanimated";
 
-import { Text } from "@/components";
+import { Text } from "@/components/ui/text";
 
 export function HelloWave() {
   const rotationAnimation = useSharedValue(0);
